refactor: combine the useCollectedProps hooks in useDrag, useDrop
