+++ packages/backend-html5/src/index.ts
@@ -2,6 +2,7 @@ import { HTML5BackendImpl } from './HTML5BackendImpl'
 import * as NativeTypes from './NativeTypes'
 import { DragDropManager, BackendFactory } from 'dnd-core'
 import { HTML5BackendContext, HTML5BackendOptions } from './types'
+export { HTML5BackendContext, HTML5BackendOptions } from './types'
 export { getEmptyImage } from './getEmptyImage'
 export { NativeTypes }
