+++ packages/react-dnd/src/hooks/useDrag/DragSourceImpl.ts
@@ -1,6 +1,6 @@
 import { invariant } from '@react-dnd/invariant'
 import { DragDropMonitor, DragSource, Identifier } from 'dnd-core'
-import { Connector } from '../../internals'
+import { SourceConnector } from '../../internals'
 import { DragSourceMonitor } from '../../types'
 import { DragObjectWithType, DragSourceHookSpec } from '../types'
 
@@ -9,7 +9,7 @@ export class DragSourceImpl<O extends DragObjectWithType, R, P>
 	public constructor(
 		public spec: DragSourceHookSpec<O, R, P>,
 		private monitor: DragSourceMonitor,
-		private connector: Connector,
+		private connector: SourceConnector,
 	) {}
 
 	public beginDrag() {
